Make sure we always add signature to `fp$` functions
Previously we were searching for the function by name but this
doesn't work when the internal name for the function is different.

In order to repro such a case the shared.c test was converted to C++
since then binaryen's internal name is different since it comes from
the de-mangled name section.